requirements.txt
Flask
numpy
pandas
pymongo
python-dotenv
scikit-learn
prophet
flask-cors
cmdstanpy
gunicorn
